… go.sum
… vendor/github.com/cloudevents/sdk-go/protocol/pubsub/v2/message.go
… vendor/github.com/cloudevents/sdk-go/v2/binding/event_message.go
… vendor/github.com/cloudevents/sdk-go/v2/binding/to_event.go
… vendor/modules.txt